改回之前的代码，测试结果误差很小，最大只有0.08

+++ debugcode/src/types.h
@@ -1,13 +1,7 @@
 #ifndef TYPES
 #define TYPES
-//#include<ap_fixed.h>
 
-//#define FXD_W_LENGTH 13
-//#define FXD_I_LENGTH 4
-
-//typedef ap_fixed<FXD_W_LENGTH,FXD_I_LENGTH> FDATA_T;
-typedef double FDATA_T;
-
+typedef float FDATA_T;
 typedef int LDATA_T;
 
 #endif
